Фильтрация. Вынос функций в отдельные файлы. Доаботка шаблонов
