+++ templates/layout.php
@@ -25,7 +25,7 @@
         <!-- PHP код для показа меню и данных пользователя -->
             <?php if ($is_auth) : ?>
                 <div class="user-menu__logged">
-                    <p><?=$user_name; ?></p>
+                    <p><?=htmlspecialchars($user_name); ?></p>
                     <a class="user-menu__bets" href="pages/my-bets.html">Мои ставки</a>
                     <a class="user-menu__logout" href="#">Выход</a>
                 </div>
@@ -55,7 +55,7 @@
             <!--список из массива категорий-->
             <?php foreach ($categories as $category): ?>
                 <li class="nav__item">
-                    <a href="pages/all-lots.html"><?=$category; ?></a>
+                    <a href="pages/all-lots.html"><?=htmlspecialchars($category); ?></a>
                 </li>
             <?php endforeach; ?>
             <!-- / список из массива категорий-->
